% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_metadata.R
\docType{data}
\name{doc_metadata}
\alias{doc_metadata}
\title{Data - Basic metadata for EEBO documents}
\format{A matrix}
\usage{
doc_metadata
}
\description{
Just ID, author, date, and title.
}
\examples{
# Some examples
}
\keyword{datasets}
