% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eebo_network.R
\docType{data}
\name{eebo_network}
\alias{eebo_network}
\title{Data - EEBO booktrade network}
\format{An igraph graph}
\usage{
eebo_network
}
\description{
Network of authors, printers, and booksellers.
}
\examples{
# Some examples
}
\keyword{datasets}
