% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/great_circle_distance.R
\name{great_circle_distance}
\alias{great_circle_distance}
\title{Computes the great-circle distance between two coordinate points.}
\usage{
great_circle_distance(lon1, lat1, lon2, lat2, radius = 3437)
}
\description{
Computes the great-circle distance between two coordinate points.
}
