% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwic.R
\docType{data}
\name{kwic}
\alias{kwic}
\title{Data - EEBO keword-in-context matrix}
\format{A matrix with 32,244 rows and 2,002 columns.}
\usage{
kwic
}
\description{
A word-context matrix drawn from the EEBO-TCP collection, built from all publicly
available files with publication years between 1640 and 1699.
}
\examples{
# Load the matrix, then analyze a term.
data(kwic)
similarity_map(mat = kwic, keyword = "rights")
}
\keyword{datasets}
