% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litmath.R
\docType{package}
\name{litmath}
\alias{litmath}
\title{Literary Mathematics: Quantitative Theory for Textual Study}
\description{
The \code{litmath} package supplments the book, Literary Mathematics,
with sample data, easy-to-use code, walkthroughs for each published
case study, and tutorials reviewing key concepts.
}
\author{
Author:
Michael Gavin


\strong{Basic operations:}
\itemize{
  \item \code{\link{similarity}}: Identifies semantic relationships among terms
  \item \code{\link{similarity_map}}: Maps conceptual structures
}

\strong{Sample Data:}
\itemize{
  \item \code{\link{kwic}}: A large matrix built from the EEBO-TCP corpus, representing
                            all texts from 1640 to 1699.
}
}
