% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_composition.R
\name{place_composition}
\alias{place_composition}
\title{Adds together the vectors corresponding to a list of place names.}
\usage{
place_composition(mat, places)
}
\description{
Adds together the vectors corresponding to a list of place names.
}
