% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_doc.R
\docType{data}
\name{place_doc}
\alias{place_doc}
\title{Data - Place Doc}
\format{A matrix}
\usage{
place_doc
}
\description{
Matrix of toponyms by documents.
}
\examples{
# Some examples
}
\keyword{datasets}
