% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_join.R
\name{place_join}
\alias{place_join}
\title{Given a gazetteer and a place name, finds all places that correspond to
and/or are contained within it.}
\usage{
place_join(places, gaz, mode = "all")
}
\description{
Given a gazetteer and a place name, finds all places that correspond to
and/or are contained within it.
}
