% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plots timeseres over a place-year matrix (1500-1699).}
\usage{
plot_timeseries(
  mat = py,
  gazetteer = gaz,
  places,
  normalize = T,
  compose_places = T,
  mode = "all"
)
}
\description{
Plots timeseres over a place-year matrix (1500-1699).
}
