% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_footprint.R
\name{semantic_footprint}
\alias{semantic_footprint}
\title{Measures the semantic footprint of a term (computed over place-keyword matrix)}
\usage{
semantic_footprint(mat, term, coords)
}
\description{
Measures the semantic footprint of a term (computed over place-keyword matrix)
}
