% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_doc.R
\docType{data}
\name{term_doc}
\alias{term_doc}
\title{Data - EEBO term-document matrix}
\format{A matrix with 2,002 rows (for each keyword) and 18,311 columns
(for each document).}
\usage{
term_doc
}
\description{
A term-document matrix drawn from the EEBO-TCP collection, built from all publicly
available files with publication years between 1640 and 1699.
}
\examples{
# Load the matrix, then analyze a term.
data(deebo)
similarity(mat = term_doc, keyword = "A48901")
}
\keyword{datasets}
